# Improve performance of generating query strings from mappings that contain sequences