--- a/CHANGES/1193.misc.rst
+++ b/CHANGES/1193.misc.rst
@@ -1,0 +1,1 @@
+Improved performance of constructing query strings from :class:`~collections.abc.Mapping` -- by :user:`bdraco`.