Roughly a 10% speed up for the tuple/list case, 5% speed up for the non sequence case

baseline

Make URL with query mapping: 0.364 sec
Make URL with query sequence mapping: 2.121 sec

this pr

Make URL with query mapping: 0.345 sec
Make URL with query sequence mapping: 1.831 sec