@bdraco 2024-10-11
```
master

Build URL with host and path and port: 0.102 sec
Build encoded URL with host and path and port: 0.068 sec
Build URL with host: 0.082 sec
Build URL with different hosts: 0.305 sec
Build URL with host and port: 0.087 sec
Make URL with host and path and port: 0.089 sec
Make encoded URL with host and path and port: 0.023 sec
Make URL with host and path: 0.090 sec
Make URL with many hosts: 0.407 sec
Make URL with IPv4 Address and path and port: 0.090 sec
Make URL with IPv4 Address and path: 0.090 sec
Make URL with IPv6 Address and path and port: 0.101 sec
Make URL with IPv6 Address and path: 0.100 sec
Make URL with query mapping: 0.348 sec
Make URL with query sequence mapping: 1.859 sec

1.9.4

Build URL with host and path and port: 0.300 sec
Build encoded URL with host and path and port: 0.072 sec
Build URL with host: 0.262 sec
Build URL with different hosts: 0.266 sec
Build URL with host and port: 0.262 sec
Make URL with host and path and port: 0.335 sec
Make encoded URL with host and path and port: 0.024 sec
Make URL with host and path: 0.312 sec
Make URL with many hosts: 0.436 sec
Make URL with IPv4 Address and path and port: 0.336 sec
Make URL with IPv4 Address and path: 0.304 sec
Make URL with IPv6 Address and path and port: 0.511 sec
Make URL with IPv6 Address and path: 0.498 sec
Make URL with query mapping: 0.346 sec
Make URL with query sequence mapping: 2.018 sec
```